Add runtime windowing system platform support to KWindowEffects
So far KWindowEffects had a compile time switch. Either it got compiled
for X11 or a dummy implementation got used. The actual runtime platform
got not considered at all. This change ensures that the X11
implementation is only used if the runtime platform is xcb. So if for
example one uses Wayland, but KWindowSystem got compiled with X11
support the dummy implementation is picked.

To achieve that a new pure virtual KWindowEffectsPrivate class is
introduced which has two implementations: KWindowEffectsPrivateDummy
and KWindowEffectsPrivateX11.

A new implementation for KWindowEffects is added which uses a global
static which holds a pointer to the KWindowEffectsPrivate. Thus the
global static functions acts as a factory method to pick the proper
implementation for the runtime system. The implementation of the
methods just delegate to the respective method of
KWindowEffectsPrivate.

REVIEW: 114922