Add unit test for KWindowInfo on X11
Unit test for most methods provided by KWindowInfo. The general pattern
is to create a window, show it, test the property, change it and
verify that the change worked. This is a little bit tricky as the test
needs to interact with large parts of the window manager. In case a
property is updated by the window manager we need to send the client
message, wait till the window manager has reacted on it and updated
the property and then wait for the property update. This is mostly done
by waiting for the signal KWindowSystem::windowChanged. Unfortunately
that reports globally and not just for the window we are interested in.
So we have to filter out till we got the correct one. If there is at
the same time further interaction with the windowing system tests can
fail, but a re-run normally fixes it.

The unit test is so far written against KWin and partially openbox,
which is used on build.kde.org. Given that KWindowInfo::actionSupported
is not tested as that is clearly to specific to the used window
manager.

REVIEW: 115190