Add platform check to KSelectionOwner and KSelectionWatcher
These are highly X11 specific classes and don't make any sense on a
platform other than xcb and were potentially crashy by using QX11Info
without verifying that the platform is used.

The implementation will now only create the d-ptr in case that the
code is running on the xcb platform and ensures that no call to xcb
is done unconditionally. All methods perform an early return in case
the d-ptr is null. Non-void methods return a sensible default value
in this case.

REVIEW: 115230