Add x11 platform detection to kkeyserver_x11
Just because it's only compiled for X11 doesn't mean it's running on
X11. So before we go into code paths which call into X11 we should
verify. That's better than:
"QX11Info::display() returns 0.  I'm probably going to crash now."

Not adjusted are obvious X11 methods which take an XLib or xcb data
type. If one calls those from non-x11 unconditionally it's an
application bug and the application deserves to crash :-P

REVIEW: 115255