Introduce QFlags for NET::Property and NET::Property2
API of KWindowInfo is adjusted to take NET::Properties and
NET::Properties2 as arguments instead of usigned long int.

The ctors of the NET* classes are adjusted which just take
NET::Properties. The other ctors take the properties array
which needs further adjustements first.

In addition a WMAllProperties and WM2AllProperties enum value
is declared to simplify matching all properties.