Remove support for _NET_WM_TAKE_ACTIVITY
This was a proposed protocol [1] which never got adopted and no toolkit
is making use of it. It's not implemented in either Qt4 or Qt5 and
also not in GTK [2].

Only KWin had support for it, but didn't announce it in NET::Supported.

This is a SIC-change, but according to lxr nothing except KWin used
these methods and there they are already in the process of getting
removed.

[1] https://mail.gnome.org/archives/wm-spec-list/2004-April/msg00013.html
[2] https://bugzilla.gnome.org/show_bug.cgi?id=154260

REVIEW: 116926