Fix crashers in KStartupInfo on remove while iterating
This is similar to the change ... just for more cases. If the API was
used with ::checkStartup it could happen that an item got removed from
the startups while iterating the list of startups. Thus the data
corrupted resulting in a crash when getting a similar crash.

In this case the code now uses erase instead of remove to have a valid
iterator.

BUG: 332681
REVIEW: 117701