Make kwindowsystemx11test more robust with openbox
The testWindowTitleChanged sometimes fails on the CI system which runs
openbox. Now an XSync is added between QWidget::setWindowTitle and
waiting for the event.

For future references of similar problems: it's required to run the
unit tests on Xvfb, running on a normal X server can result in different
behavior.