Port KWindowSystem::raiseWindow and ::lowerWindow to XCB
No unit test as that is extremely difficult to test as it depends
on what the window manager running in the test environment supports.
The code supports two variants for whether the WM supports the feature
or not, thus the test would need to change the WM during execution and
this would affect other running tests.

REVIEW: 118253