Improve CMake error message when XRender or XFixes are missing
This is what would be printed previously:

CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
X11_Xrender_LIB (ADVANCED)
    linked by target "KF5WindowSystem" in directory /foo/bar

Now we print:

The XRender library could not be found. Please install the development package for it.

REVIEW: 119849