Add support for input hint in NETWinInfo
The Input Hint is part of WM_HINTS property which is already processed
for WM2GroupLeader and WM2Urgency.

The default for input hint is true. ICCCM suggests to assume convenient
values for all fields of WM_HINTS property if a window is mapped wihtout
one. KWin sets this value to true, so the sane convenient value for the
input field seems to be true.

REVIEW: 120165
CHANGELOG: NETWinInfo provides convenient wrapper for input hint of WM_HINTS property.