Add test case for NET::DemandsAttention to KWindowInfoX11Test
The test is a little bit tricky as there are differences in the
behavior of various window managers. For example KWin's focus
stealing prevention kicks in when the test window is shown and marks
it as demands attention, while openbox (CI system) focuses directly
and sets no demands attention. The test circumvents this problem by
force active window on it.

In addition to testing the EWMH property also the urgency hint from
ICCCM is considered. Some window managers (at least KWin and openbox)
map this hint to demands attention.