Try to make KWindowInfoX11Test::testDemandsAttention more stable on CI
Unfortunately the test is currently failing on the CI system for to me
unknown reasons. On a similar local setup it has never failed at that
point.

What I try with this change is to not wait by time till our test window
gets focused, but use a change signal. Also it improves the KWin focus
stealing prevention workaround.

Obviously the test passes locally with both KWin and openbox. If it
still fails on the CI system after this change I'm out of ideas and
will disable the test for openbox.