Try flushing xcb connection in KWindowInfoX11Test::testDemandsAttention
Maybe that makes the test pass on the CI system. The thought is that
the test fails when waiting for the property state change which should
be set by the WM. As there is enough time (500 msec) for the WM to do
that, it's more likely that the event never reaches the WM. We had
problems with that in the past, that Qt didn't flush the connection.