Drop incorrect warnings when using KXMessages without QX11Info
The idea was to not perform the action if QX11Info reports that its
not the xcb plugin. But this check is not correct as those methods
are not going through QX11Info at all. In addition they pass in either
an xcb connection or XLib display which is totally fine and needed
for example if we want Wayland applications interact with the X11
world. Also it causes problems for e.g. kdeinit which is not a
Q*Application and thus does not have a QX11Info.

At the same time fix an incorrect usage of QX11Info::connection
where an xcb_connection_t* is already passed in to the method.

BUG: 340310
REVIEW: 121198