Add new overload to KWindowSystem::icon
The new overload is a special solution for the X11 platform. It takes
the NETWinInfo as argument to read the information from.

This significantly reduces the time to fetch icons for users who
already have a NETWinInfo with the required data. E.g. for the case
of KWin it can reduce the number of roundtrips to the X Server from
up to 15 to 0.

For non-X11 platforms the method just delegates to the method not
taking the NETWinInfo argument.

CHANGELOG: New overload to KWindowSystem::icon which reduces roundtrips to X-Server
REVIEW: 122144