Do not require KSelectionWatcher and KSelectionOwner to be on platform xcb
Adding dedicated constructor overloads taking the xcb_connection_t* and
the root window. If these overloads are used it is not verified that
platform xcb is used and QX11Info is not used. The other overloads still
verify that platform is xcb and depend on QX11Info.

REVIEW: 122604
CHANGELOG: KSelectionWatcher and KSelectionOwner can be used without depending on QX11Info.