Better support for non-platform xcb in KXMessages
The static methods of KXMessages already support being used without
platform xcb (and that way without a QGuiApplication). This change
makes it possible to also use KXMessages without platform xcb. A new
ctor overload is added which accepts the xcb_connection_t* and the
root window to be used. The existing ctor still uses QX11Info and
verifies that it is used on platformX11.

REVIEW: 122607
CHANGELOG: KXMessages can be used without depending on QX11Info