Support KWindowSystem::icon with NETWinInfo on all platforms
If we compile with X11 support we want this method to use the X11
implementation on all platforms and not just on platform xcb. The
NETWinInfo is already so X11 specific that the platform doesn't
matter.

To solve this the method it delegates to in KWindowSystemPrivateX11
is made static. It only operates through the NETWinInfo and doesn't
need anything else from KWindowSystemPrivateX11.

REVIEW: 122770
CHANGELOG: KWindowSystem::icon with NETWinInfo not bound to platform X11.