[autotests] Drop incorrect testcase in KWindowSystemX11Test
We tried to get the signal
KWindowSystem::windowChanged(WId id, const unsigned long *properties)
and store it in a QSignalSpy. But that could not work as the array
for the signal gets created on the stack. Thus the pointer stored in
the signalspy points to deleted memory. We can read it but it certainly
is not the value we expect. That this used to work was poor chance.