[autotests] Try to fix KWindowInfoX11Test::testGeometry
The test for geometry started to fail reliable on the new CI system,
but I'm unable to reproduce on any system.

Given the test output it looks like the window manager places the
window. The expected geometry is QRect(0,0 640x480), but the actual
geometry is QRect(480,368 640x480). This means the size is correct,
but the position isn't, which sounds a lot like interfering window
manager.

Thus the test is adjusted to only check the initial size instead of
geometry. The other test to change is kept. It's quite likely that
the test will now fail in them. If that's the case I'm just going to
drop the test case.