Change source file layout
In the src directory we introduce a platforms directory with a specific
* osx
* wayland
* windows
* xcb

sub directory. The platform specific source and header files are moved
into those directories.

This is a preparation step to move the platform specific behavior into
plugins which get loaded at runtime. Such a change is required to
support more platforms in future for which kwindowsystem cannot provide
support directly. An example is proper Wayland support on the Plasma
platform. It needs to provide its own implementationn which has to
differ from the generic Wayland implementation.

REVIEW: 124212