Handle clients having a different device pixel ratio
Qt scales down the size of QScreens by the device pixel ratio; we should
make our windowing APIs match.

X (and KWin) deal with device dependent pixels, so everything needs to
be converted when communitcating geometry.

Abstraction happens in the main kwindowsystem so X and Wayland are both
supported.

BUG: 350865
BUG: 350614
BUG: 347951
REVIEW: 124648