[idletime] Tear-down Wayland objects before app quits
The created Wayland objects were destroyed when the Poller got
destroyed. In the case of KIdleTime the Poller gets destroyed on library
unload which happens after the QGuiApplication was destroyed. This means
the Wayland connection used to create the objects does not exist any
more.

With latest Wayland library trying to destroy objects after the Wayland
connection got destroyed results in a crash as we can see currently on
build.kde.org.

This change implements custom Wayland cleanup code in
Poller::unloadPoller. To ensure that unloadPoller gets invoked a
connection on destroy of platformNativeInterface is set up which happens
before the Wayland connection is getting destroyed.

This change hopefully fixes the crash on build.kde.org. It might be that
there is another object not yet destroyed, we'll see.