[autotests] Fix broken KWindowInfoTest
The test started to break with the last change due to the test being
wrong. It still operated on the deprecated windowChanged signals and
casted a NET::Property2 into a NET::Property and that into a uint. It's
not surprising that this can fail.

The test now operates on the windowChanged signal which has
NET::Properties and NET::Properties2 as arguments.