Support flatpak platform
Summary:
Flatpak isn't a real platform, but uses either "xcb" or "wayland". Thus
we need to detect which one it is. With D8223 the flatpak platform sets
the "real" platform as an env variable. Thus we can make use of it to
detect the actual used platform and provide the proper platform.

Reviewers: #frameworks, #flatpak, apol

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8224