[xcb] Fix implementation of _NET_WM_FULLSCREEN_MONITORS
Summary:
According to NETWM spec the client "wishing to change this list MUST send
a _NET_WM_FULLSCREEN_MONITORS client message to the root window". This
was not implemented at all, instead the property was updated. The
property must be set by the window manager, or as the spec says:
"The Window Manager MUST keep this list updated to reflect the current
state of the window."

In KWin (as the user of NETWinInfo) this was implemented correctly (see
kwin, file netinfo.cpp method WinInfo::changeFullscreenMonitors).

BUG: 391960

Test Plan:
New test case added which verifies the client/wm interplay,
old and incorrect test case removed

Reviewers: #frameworks, #kwin, #plasma

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11414