Span wayland objects to lifespan of the QApplication
Summary:
We must release all wayland objects before the QPA connection is closed.

We used to do this explicitly, but this still left an awkward code path
where we could still try to recreate a BlurManager if it's called after
the QApplication is closed.

Instead we can scope all the wayland objects to the QApplication. The
objects themselves were not leaked as public API so this is safe. Calls
after this will simply no-op.

This fixes crashes on tear down of plasma.

BUG: 372789

Test Plan: kquitapp5 plasmashell

Reviewers: #plasma

Subscribers: apol, ngraham, anthonyfieroni, zzag, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22084