[platforms/xcb] Use XRES extension to get real window PID
Summary:
It's increasingly common for apps to be in their own PID namespace. If
this is the case, they report a buggy NET_WM_PID on their windows as
obviously they don't know their own PID.

This patch uses the XResources extension to query the real PID instead
of asking the window.

This allows the task manager to know the real PID when doing lookups for
matching services or showing pulseaudio badges.

Querying the NET_WM_PID with the NETWM wrappers remains unchanged, so we
still have access to all information should a system need it.

BUG: 384837

Test Plan:
The relevant unit test KWindowInfo::testPid() still passes whilst using
this new extension.

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22755