KWindowSystem: add cmake option KWINDOWSYSTEM_NO_WIDGETS
Summary:
This should obviously not be set by distributors, but it can be used
for testing, for porting (code that uses this library) and for special
deployments (e.g. Android)

Test Plan:
Builds (on top of the other kwindowsystem change requests).
Breaks other modules of course, at this point, I'm only starting to port them.

Reviewers: broulik, davidedmundson, zzag, nicolasfella

Reviewed By: zzag

Differential Revision: https://phabricator.kde.org/D23216