Introduce shadows API
Summary:
The new API provides a platform-independent way for setting drop-shadows
on a window. X11 and Wayland details are completely hidden from the users
of the new API, but you still need to be cautious about QPAs that destroy
the underlying native resources(e.g. wl_surface) upon a window becoming
hidden. It is highly recommended to install an event filter that reacts
to events of type QEvent::PlatformSurface.

Reviewers: #kwin, davidedmundson

Reviewed By: #kwin, davidedmundson

Subscribers: davidedmundson, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T12496

Differential Revision: https://phabricator.kde.org/D26457