[KWindowShadows] Check for X connection
On teardown of plasmashell the shared pointer with shadows is unref'd quite late (since PanelShadows is a singleton)
at which point QX11Info::connection() already return null, leading to a crash in those xcb calls.
The old code in Plasma PanelShadows checked for QX11Info::display() being null, which is replicated here.

Differential Revision: https://phabricator.kde.org/D26960