[xcb] Send correctly scaled icon geometry
Otherwise when using Qt scaling on X11, icon information has the wrong
size and location on screen, with the user-visible effect that window
Minimize animations minimize the window to the wrong location when
using Qt scaling on X11 with a high DPI screen. The effect is especially
hilarious with the Magic Lamp minimize effect and a bottom panel. This
change fixes all of that so that the minimize animations work properly
on X11 with Qt scaling and a high DPI screen.

BUG: 407458
FIXED-IN: 5.72