Fix blur/contrast region not getting updated
There is a misunderstanding of current blur protocol in the previous fix.
The update{Blur/Contrast}Region is only triggered by setting a new blur
object, which essentially ignore update the future update of blur
region.

This tries to restore the old behavior, by always creating a new blur
object, but always make sure we destroy them when being replaced.

BUG: 456442