Remove QWindow::isExposed() check in activateWindow()
A window can be considered as unexposed if it's not visible, e.g. it's
minimized or not on current virtual desktop. The isExposed() check
neuters the desired side-effects of the activateWindow() request.

As far as I understand, QWindow::isExposed() check was added to ensure
that there's wl_surface, but kwayland-integration noops if the
wl_surface is missing.

BUG: 458983
Co-authored: Nicolas Fella <nicolas.fella@gmx.de>