Suppress "could not create KWindowShadow" warnings in plasmashell
Plasma::Dialog can attempt to create a window shadow before the window
becomes visible on the screen. This is a legit case but WindowShadow
will print a warning if that happens.

This change rearranges the creation logic a bit. If the window already
has a surface, KWindowShadow::create() will succeed. If the window does
not have a surface, an event filter will be installed and the shadow is
going to be created when the window is exposed. On the other hand, if
the shadow manager global is not present, KWindowShadow::create() will
return false.