Reimplement shmpool
Because KWayland doesn't handle compositor restarts, we would
crash with the parent commit only when creating a new shadow tile
because of using a wl_buffer from the old wl_display as an argument
in a request to the new one. The implementation is lifted out of
KWayland and simplified a bit as it only needs to cover the
exact use case needed here.
When the compositor is restarted all current buffers are deleted,
if a shadow tile is re-used with a new window or a current window
receives an expose event the buffer will be recreated.