wayland: Avoid creating the platform window in surfacehelper
All window effect code has code to dynamically handle the window
being created and torn down to update the wayland state. We
don't need or want to explictly create a window ahead of time.

This is proving problematic as there is a code-path that causes
us to re-enter QWindow::create unsafely.

CCBUG: 499383