windowshadow: Fix potential crash on shutdown
ShadowManager is created with a parent, QGuiApplication.

It can be accessed whilst deleting other children of QGuiApplication. If
it accessed after it is torn down, we have a dangling pointer.

Instead guard access to the singleton based on the availability of the
QGuiApplication. This matches what we do in KGuiAddons.

BUG: 487660